fix(menubar): delete the full-sort path that rearranges the whole bar
