+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,8 +1,7 @@
 name: Bug Report
 description: File a bug report
 title: "[Bug] "
-type: bug
-projects: ["thaw-app/1"]
+labels: ["bug"]
 body:
   - type: markdown
     attributes:
