+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,13 +1,12 @@
 name: Feature Request
 description: Request a feature or suggest an idea.
 title: "[Feature Request] "
-type: feature
-projects: ["thaw-app/1"]
+labels: ["feature"]
 body:
   - type: markdown
     attributes:
       value: |
-        **Before submitting:** Search [open & closed suggestions](https://github.com/thaw-app/Thaw/issues?q=is%3Aissue+-type%3ABug) for duplicates.
+        **Before submitting:** Search [open & closed suggestions](https://github.com/thaw-app/Thaw/issues?q=is%3Aissue+-label%3Abug) for duplicates.
 
         > [!IMPORTANT]
         > macOS 27 / Golden Gate support and preview builds are tracked in **[#687](https://github.com/thaw-app/Thaw/issues/687)**. Do not open a new feature request that is just “support macOS 27”.
