+++ .github/workflows/ci.yml
@@ -102,7 +102,7 @@ jobs:
 
   build:
     if: github.repository == 'thaw-app/Thaw'
-    runs-on: xcode-27
+    runs-on: macos-26
     needs: [changes, static_analysis]
     env:
       HOMEBREW_NO_AUTO_UPDATE: 1
@@ -113,6 +113,9 @@ jobs:
 
       - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
         if: needs.changes.outputs.code == 'true'
+      - run: sudo xcode-select -s /Applications/Xcode_26.5.app/Contents/Developer
+        if: needs.changes.outputs.code == 'true'
+
       - name: Cache SwiftPM
         if: needs.changes.outputs.code == 'true'
         uses: actions/cache@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
