+++ .github/workflows/release.yml
@@ -77,7 +77,7 @@ permissions:
 jobs:
   release:
     if: github.repository == 'thaw-app/Thaw'
-    runs-on: blacksmith-6vcpu-macos-26
+    runs-on: macos-26
     # Trust boundary: required reviewers on the `release` Environment
     # (diazdesandi). Other org owners can still unblock via admin bypass.
     # Secrets are not consumed until approval.
@@ -679,7 +679,7 @@ jobs:
   attach-provenance:
     needs: [release, attest-dmg, attest-sbom]
     if: ${{ github.repository == 'thaw-app/Thaw' && !inputs.dry_run }}
-    runs-on: blacksmith-4vcpu-ubuntu-2404
+    runs-on: ubuntu-latest
     permissions:
       contents: write
     steps:
