+++ .github/workflows/stale.yml
@@ -25,4 +25,4 @@ jobs:
           stale-issue-label: "stale"
           stale-issue-message: "This issue has been automatically marked as stale due to inactivity."
           close-issue-message: "Closing this issue due to prolonged inactivity."
-          exempt-issue-labels: "help wanted,keep open"
+          exempt-issue-labels: "help wanted,keep open,feature,enhancement"
