+++ ThawTests/Settings/Models/AdvancedSettingsSnapshotTests.swift
@@ -43,7 +43,6 @@ struct AdvancedSettingsSnapshotTests {
             enableDiagnosticLogging: false,
             useDoubleClickToShowAlwaysHiddenSection: false,
             useOptionClickToShowAlwaysHiddenSection: false,
-            useLCSSortingOnNotchedDisplays: false,
             enableMenuBarItemOverflow: false,
             searchSectionOrder: ["visible", "hidden", "alwaysHidden"],
             searchIncludeVisible: true,
@@ -67,7 +66,6 @@ struct AdvancedSettingsSnapshotTests {
             enableDiagnosticLogging: true,
             useDoubleClickToShowAlwaysHiddenSection: true,
             useOptionClickToShowAlwaysHiddenSection: true,
-            useLCSSortingOnNotchedDisplays: true,
             enableMenuBarItemOverflow: true,
             searchSectionOrder: ["alwaysHidden", "hidden", "visible"],
             searchIncludeVisible: false,
@@ -326,7 +324,6 @@ struct AdvancedSettingsSnapshotTests {
             enableDiagnosticLogging: false,
             useDoubleClickToShowAlwaysHiddenSection: false,
             useOptionClickToShowAlwaysHiddenSection: false,
-            useLCSSortingOnNotchedDisplays: false,
             enableMenuBarItemOverflow: false,
             searchSectionOrder: ["visible", "hidden", "alwaysHidden"],
             searchIncludeVisible: false,
@@ -366,7 +363,6 @@ struct AdvancedSettingsSnapshotTests {
             enableDiagnosticLogging: true,
             useDoubleClickToShowAlwaysHiddenSection: true,
             useOptionClickToShowAlwaysHiddenSection: true,
-            useLCSSortingOnNotchedDisplays: true,
             enableMenuBarItemOverflow: true,
             searchSectionOrder: ["visible", "hidden", "alwaysHidden"],
             searchIncludeVisible: true,
