+++ sonar-project.properties
@@ -35,17 +35,6 @@ sonar.exclusions = \
 # all of which stay measured and carry dedicated suites in ThawTests. New
 # decision logic belongs in those files, not back in the manager.
 
-# ProfileManager.swift and DisplaySettingsManager.swift were split the same
-# way: the measured files keep the manifest/file CRUD, capture, persistence,
-# lookup, URI handling and every decision rule, while the +Live companions
-# hold the halves that need a running AppState — pushing snapshots into live
-# managers, the spacing relaunch wave, NSScreen/WindowServer display state,
-# Carbon hotkey registration, Focus Filter intents, and an app-modal alert.
-# Capture now takes the three stores it actually reads (settings, appearance
-# manager, item manager) instead of the full AppState, so save/update paths
-# stay measured; the AppState overloads forwarding to them live in the
-# excluded companions. New decision logic belongs in the measured files.
-
 # MenuBarSection.swift and MenuBarItem.swift were split the same way rather
 # than excluded whole, because both buried pure logic that the rule above says
 # must stay measured.
@@ -110,8 +99,6 @@ sonar.coverage.exclusions = \
   Thaw/Permissions/PermissionsView.swift,\
   Thaw/Permissions/PermissionsWindow.swift,\
   Thaw/Settings/Models/AppSettings.swift,\
-  Thaw/Settings/Models/ProfileManager+Live.swift,\
-  Thaw/Settings/Models/DisplaySettingsManager+Live.swift,\
   Thaw/Settings/SettingsPanes/*.swift,\
   Thaw/Settings/SettingsDetailStyle.swift,\
   Thaw/Settings/SettingsView.swift,\
